Several bug fixes and improvements:
~ Clean up core patches initializer
~ Add theming bug fix to core, plugin is no longer required (hopefully)
~ Handle the fact that name in json and file name may not match
To handle this, I put the offset "name" prop into "displayName" and then set the filename to be the same as "name".
This means that the theme can still display a prettier name but will work properly to apply and uninstall, unlike before.
This cannot be handled on the JS side any other way because the file names of the native side are not exposed on the JS side, so the only solution is to do it during installation.
~ Add new utility method: writeFile, takes in a path where "test.json" is actually Discord.app/Documents/test.json, and content, as a string